Rise and Swap Grid System (12/24/2025)
This is the grid system for my game. The only problem I have with it is that I can not see the grid when I run the system. The grid is made with the right colors just to test it out, but the camera is not able to focus on the grid the way I want it to and I don't know why. If anyone can help me with that, that would be great!